Isolate connection handshake methods
