+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -1,1 +1,2 @@
+pub mod connection;
 pub mod update_client;
